accc: fix some visual studio warnings
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@7725 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: ef75e4ccbe641bb8acf71d43f4ca0cb306d5606e
Former-commit-id: ee5213096837c1b7449f6785aad3ab6328d2806f

+++ src/tools/accc/ac3dload.cpp
@@ -201,7 +201,7 @@ void obSetVertexArraysIndex(ob_t * ob, int vaIdx, int newIndex)
 #ifndef M_PI
 #define M_PI 3.14159267
 #endif
-extern int far_dist;
+extern double far_dist;
 void computeTriNorm(ob_t * object);
 void smoothTriNorm(ob_t * object);
 void computeObjectTriNorm(ob_t * object);
